Asynchronous decryption fix
RNCryptor uses dispatch_get_current_queue() in initWithHandler: as the responseQueue. However, due to the recent deprecation of dispatch_get_current_queue() in iOS6 this leads to a situation where the handler is never called. This commit changes the initWithHandler to create a new queue that is then used as the response queue which allows asynchronous decryption to work correctly.

In addition to the change it includes a unit test to verify that it works correctly.

This commit fixes #43

Signed-off-by: Calman Steynberg <calman@steynberg.ca>