import modular CommonCrypto header if available
With XCode 10 and swift 4.2 the CommonCrypto library seems to have gotten the long awaited modular header. So if it is available we do not need the Objective-C Bridging header anymore. It suffices to import the RNCryptor.swift file in a project. It is even possible to paste the content on RNCryptor.swift into a swift playground on the iPad and you can then use all the Encryption/Decryption functionalities in the Playground.